Stop using select.poll .
It was mostly to train this code, but reduces portability. As it is not
really needed (no other descriptor to poll), use to handleEventsTimeout
instead. This fixes long stalls (for some reason, transfer timeouts are not
always honoured, leading to long time without data when there are too few
events to capture).