parser: Improve incomplete transaction detection.
When encountering an invalid token at the beginning of a transaction,
the previous transaction would get marked as incomplete, which is
incorrect. Instead, trigger an EOF token to try to reduce the grammar.
If the error does happen on an EOF token, it means current started
transaction is really incomplete, so push it as such.